refactor(web): isolate session recovery state
